Make EnvoyProxyMiddleware a Gin-style middleware
