+++ go.mod
@@ -4,6 +4,7 @@ go 1.15
 
 require (
 	github.com/dgrijalva/jwt-go v3.2.0+incompatible
+	github.com/gin-gonic/gin v1.6.3
 	github.com/golang/gddo v0.0.0-20201222204913-17b648fae295
 	github.com/google/go-cmp v0.3.1 // indirect
 	go.uber.org/zap v1.16.0
